Temporarily remove internal link references
NR Employee here :)
I met with @barbnewrelic and we discussed removing the links referencing other NR docs, but leaving the links to external sources and docs (Fluent Bit, VS code stuff, etc.). The goal is to "start fresh" and add links in again later as necessary. I'm working on some additional formatting/content changes and updates but trying to do so incrementally as Barb recommended.